+++ projects/ng-dynamic-json-form/src/lib/models/extra/form-control-extra-date.model.ts
@@ -1,6 +1,7 @@
 export interface FormControlExtraDate {
   min?: string;
   max?: string;
+  selectTime?: boolean;
   displayFormat?: string;
   outputFormat?: string;
 }