+++ projects/ng-dynamic-json-form/ui-material/lib/ui-material-input/ui-material-input.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, inject } from '@angular/core';
+import { Component } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatInputModule } from '@angular/material/input';
@@ -7,8 +7,6 @@ import {
   ErrorMessageService,
   NgDynamicJsonFormCustomComponent,
 } from 'ng-dynamic-json-form';
-import { Observable } from 'rxjs/internal/Observable';
-import { debounceTime, startWith, switchMap } from 'rxjs/operators';
 
 @Component({
   selector: 'ui-material-input',
@@ -18,21 +16,4 @@ import { debounceTime, startWith, switchMap } from 'rxjs/operators';
   styles: [],
   providers: [ErrorMessageService],
 })
-export class UiMaterialInputComponent extends NgDynamicJsonFormCustomComponent {
-  errors$?: Observable<string[]>;
-
-  private errorMessageService = inject(ErrorMessageService);
-
-  ngOnInit(): void {
-    this.errors$ = this.control?.valueChanges.pipe(
-      startWith(this.control.value),
-      debounceTime(0),
-      switchMap(() =>
-        this.errorMessageService.getErrors$(
-          this.control!,
-          this.data?.validators || []
-        )
-      )
-    );
-  }
-}
+export class UiMaterialInputComponent extends NgDynamicJsonFormCustomComponent {}