+++ projects/ng-dynamic-json-form/ui-material/lib/ui-material-switch/ui-material-switch.component.ts
@@ -3,11 +3,17 @@ import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NgDynamicJsonFormCustomComponent } from 'ng-dynamic-json-form';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { MatInputModule } from '@angular/material/input';
 
 @Component({
   selector: 'ui-material-switch',
   standalone: true,
-  imports: [CommonModule, ReactiveFormsModule, MatSlideToggleModule],
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    MatSlideToggleModule,
+    MatInputModule,
+  ],
   templateUrl: './ui-material-switch.component.html',
   styles: [],
 })